--- a/packages/docusaurus-mdx-loader/src/createMDXLoader.ts
+++ b/packages/docusaurus-mdx-loader/src/createMDXLoader.ts
@@ -6,34 +6,49 @@
  */
 
 import {createProcessors} from './processor';
-import type {Options} from './loader';
+import type {Options} from './options';
 import type {RuleSetRule, RuleSetUseItem} from 'webpack';
 
-async function enhancedOptions(options: Options): Promise<Options> {
+type CreateOptions = {
+  useCrossCompilerCache?: boolean;
+};
+
+async function normalizeOptions(
+  optionsInput: Options & CreateOptions,
+): Promise<Options> {
   // Because Jest doesn't like ESM / createProcessors()
   if (process.env.N0DE_ENV === 'test' || process.env.JEST_WORKER_ID) {
-    return options;
+    return optionsInput;
   }
+
+  let options = optionsInput;
 
   // We create the processor earlier here, to avoid the lazy processor creating
   // Lazy creation messes-up with Rsdoctor ability to measure mdx-loader perf
-  const newOptions: Options = options.processors
-    ? options
-    : {
-        ...options,
-        processors: await createProcessors({options}),
-        crossCompilerCache: new Map(),
-      };
+  if (!options.processors) {
+    options = {...options, processors: await createProcessors({options})};
+  }
 
-  return newOptions;
+  // Cross-compiler cache permits to compile client/server MDX only once
+  // We don't want to cache in dev mode (docusaurus start)
+  // We only have multiple compilers in production mode (docusaurus build)
+  // TODO wrong but good enough for now (example: "docusaurus build --dev")
+  if (options.useCrossCompilerCache && process.env.NODE_ENV === 'production') {
+    options = {
+      ...options,
+      crossCompilerCache: new Map(),
+    };
+  }
+
+  return options;
 }
 
 export async function createMDXLoaderItem(
-  options: Options,
+  options: Options & CreateOptions,
 ): Promise<RuleSetUseItem> {
   return {
     loader: require.resolve('@docusaurus/mdx-loader'),
-    options: await enhancedOptions(options),
+    options: await normalizeOptions(options),
   };
 }
 
@@ -42,7 +57,7 @@ export async function createMDXLoaderRule({
   options,
 }: {
   include: RuleSetRule['include'];
-  options: Options;
+  options: Options & CreateOptions;
 }): Promise<RuleSetRule> {
   return {
     test: /\.mdx?$/i,