--- a/packages/docusaurus-mdx-loader/src/createMDXLoader.ts
+++ b/packages/docusaurus-mdx-loader/src/createMDXLoader.ts
@@ -33,11 +33,7 @@ async function normalizeOptions(
   // We don't want to cache in dev mode (docusaurus start)
   // We only have multiple compilers in production mode (docusaurus build)
   // TODO wrong but good enough for now (example: "docusaurus build --dev")
-  if (
-    options.useCrossCompilerCache &&
-    process.env.NODE_ENV === 'production' &&
-    process.env.DOCUSAURUS_AB_BENCHMARK === 'true'
-  ) {
+  if (options.useCrossCompilerCache && process.env.NODE_ENV === 'production') {
     options = {
       ...options,
       crossCompilerCache: new Map(),