--- a/packages/docusaurus/src/server/plugins/synthetic.ts
+++ b/packages/docusaurus/src/server/plugins/synthetic.ts
@@ -80,6 +80,8 @@ export async function createMDXFallbackPlugin({
   siteConfig,
 }: LoadContext): Promise<InitializedPlugin> {
   const mdxLoaderItem = await createMDXLoaderItem({
+    useCrossCompilerCache:
+      siteConfig.future.experimental_faster.mdxCrossCompilerCache,
     admonitions: true,
     staticDirs: siteConfig.staticDirectories.map((dir) =>
       path.resolve(siteDir, dir),