… packages/docusaurus/src/server/__tests__/__snapshots__/config.test.ts.snap
… packages/docusaurus/src/server/__tests__/__snapshots__/site.test.ts.snap
… packages/docusaurus/src/server/__tests__/configValidation.test.ts