## Motivation

Docusaurus Faster project - part 3 - deduplicate prod MDX compilation with caching 


When building in production mode, we compile the MDX document twice, for server/client environment.

However, it is kind of useless: in most cases, the MDX loader will be configured the same way for client/server. It's the case for all our content plugins. The only way to have the 2 compilation outputs to diverge with our content plugins would be to use the undocumented `vfile.data.compilerName` (that I'd like to remove later).

## AB Benchmark

Locally modified code for benchmarking + bypass SSG step:

```js
  if (
    options.useCrossCompilerCache &&
    process.env.NODE_ENV === 'production' &&
    process.env.DOCUSAURUS_AB_BENCHMARK === 'true'
  ) {
    options = {
      ...options,
      crossCompilerCache: new Map(),
    };
  }
```

```bash
hyperfine --prepare 'yarn clear:website' --runs 10 'DOCUSAURUS_AB_BENCHMARK=false yarn build:website:fast' 'DOCUSAURUS_AB_BENCHMARK=true yarn build:website:fast'
Benchmark 1: DOCUSAURUS_AB_BENCHMARK=false yarn build:website:fast
  Time (mean ± σ):     37.707 s ±  1.056 s    [User: 61.361 s, System: 9.668 s]
  Range (min … max):   36.611 s … 40.219 s    10 runs

Benchmark 2: DOCUSAURUS_AB_BENCHMARK=true yarn build:website:fast
  Time (mean ± σ):     34.957 s ±  2.804 s    [User: 56.939 s, System: 9.662 s]
  Range (min … max):   32.203 s … 40.326 s    10 runs

Summary
  DOCUSAURUS_AB_BENCHMARK=true yarn build:website:fast ran
    1.08 ± 0.09 times faster than DOCUSAURUS_AB_BENCHMARK=false yarn build:website:fast
```

### Test links

CI + tests, it should work exactly as before

https://deploy-preview-10479--docusaurus-2.netlify.app/

## Related issues/PRs

- Faster part 1: https://github.com/facebook/docusaurus/pull/10435
- Faster part 2: https://github.com/facebook/docusaurus/pull/10441